requirements.txt

fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic-settings==2.1.0
pinecone-client==3.0.0
sentence-transformers==2.3.1
groq==0.5.0
PyPDF2==3.0.1
slowapi==0.1.9
python-dotenv==1.0.0
numpy
